requirements.txt
Flask==1.1.2
requests==2.25.1
Flask-Cors==3.0.10
Jinja2==2.11.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
Werkzeug==1.0.1
python-dotenv==0.15.0